more mock api tests
