+++ src/Utils/Utils.csproj
@@ -49,6 +49,7 @@
       <Link>Properties\SharedAssemblyInfo.cs</Link>
     </Compile>
     <Compile Include="GeoCoordinate.cs" />
+    <Compile Include="UnixTimestampUtils.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Constants.cs" />
     <Compile Include="GeoCoordinateBounds.cs" />
