+++ src/Utils/UtilsUWP.csproj
@@ -77,6 +77,7 @@
     <Compile Include="JsonConverters\PolylineToGeoCoordinateListConverter.cs" />
     <Compile Include="JsonConverters\JsonConverters.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="UnixTimestampUtils.cs" />
     <Compile Include="Vector2d\Mathd.cs" />
     <Compile Include="Vector2d\RectD.cs" />
     <Compile Include="Vector2d\Vector2d.cs" />
